Move network badge
